+++ src/modules/filter/store/stateFilter.js
@@ -32,6 +32,7 @@
  * @property {Boolean} closeGfi - If it is true, the open gfi window will be closed after new filtering.
  * @property {Boolean} preventAdjust - If it is true, the snippet will not be adjusted.
  * @property {String} questionLink - The url of the question link.
+ * @property {Boolean} closeDropdownOnSelect - if true, the dropdown list will be closed after choosing option on the multiselect dropdown.
  */
 
 const state = {
@@ -66,7 +67,8 @@ const state = {
     preventAdjust: false,
     linkText: "",
     initialStartupReset: false,
-    questionLink: ""
+    questionLink: "",
+    closeDropdownOnSelect: true
 };
 
 export default state;